System now has isolated PCI transfer format from the output format of CP2exec.
Note that the datagram formatting out of
CP2exec is completely unworkable. The receiving
programs, QtDsp and then CP2Scope, index into the
datastructures using numeric offsets. This
will be completely revamped, but for the moment
it needs to be considered if trying to make 
any changes to the datagram protocol.